test: add fsx_paranoid CAS round-trip variant
Restore main's tests/fsx.rs (canonical xfstests fsx binary, 50k random ops)
and pull in the paranoid mode from PR #41's earlier fsx implementation as a
separate test file.

Paranoid mode does full CAS round-trip after each mutation: write/truncate,
flush + sleep, re-open + read-back, verify against in-memory reference. This
catches composition bugs in range_upload that the canonical fsx misses (since
fsx reads from the local staging file, not from CAS).

Slow (~1.5s per op for flush debounce + CAS propagation). Use FSX_PARANOID_OPS
to control iteration count (default: 100).
